Restructuring-code
