+++ scripts/RapidAPI/README.md
